requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy[asyncio]>=1.5.0
asyncpg>=0.27.0
authlib>=1.0.0
python-dotenv>=1.0.0
pyjwt>=2.0.0
passlib[bcrypt]>=1.7.0
httpx>=0.24.0
python-multipart>=0.0.6
email-validator>=1.1.0
openai>=1.0.0 ; extra == 'openai'

# Optional (Langraph SDK - replace with real package name if different)
# langraph
